Separate out ui components to make it a bit easier to view and manage
